Prepare cross platform GitCommander class
